# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Polman API","version":"0.1.0"},"servers":[{"url":"http://localhost:8000/polman","description":"Local Development"},{"url":"http://10.160.3.20:30700/polman","description":"ICOS Staging Testbed"}],"paths":{"/registry/api/v1/policies/{id}":{"get":{"tags":["policies","policies"],"summary":"Get Policy","operationId":"get_policy_registry_api_v1_policies__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PolicyRead"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/registry/api/v1/policies/":{"get":{"tags":["policies","policies"],"summary":"List Policies","operationId":"list_policies_registry_api_v1_policies__get","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PolicyRead"},"title":"Response List Policies Registry Api V1 Policies  Get"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["policies","policies"],"summary":"Create Policy","operationId":"create_policy_registry_api_v1_policies__post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PolicyCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PolicyRead"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/registry/api/v1/icos/":{"post":{"tags":["policies","icos"],"summary":"Icos Process App Descriptor","operationId":"icos_process_app_descriptor_registry_api_v1_icos__post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ICOSAppDeployedRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PolicyRead"},"title":"Response Icos Process App Descriptor Registry Api V1 Icos  Post"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/watcher/api/v1/webhooks/alertmanager":{"post":{"tags":["webhooks","webhooks"],"summary":"Alertmanager Webhook","operationId":"alertmanager_webhook_watcher_api_v1_webhooks_alertmanager_post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlertmanagerWebhook"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/status/":{"get":{"tags":["status"],"summary":"Read Item","operationId":"read_item_status__get","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AlertmanagerAlert":{"properties":{"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"},"labels":{"additionalProperties":{"type":"string"},"type":"object","title":"Labels"},"annotations":{"additionalProperties":{"type":"string"},"type":"object","title":"Annotations"},"startsAt":{"type":"string","format":"date-time","title":"Startsat"},"endsAt":{"type":"string","format":"date-time","title":"Endsat"},"generatorURL":{"type":"string","title":"Generatorurl"},"fingerprint":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Fingerprint"}},"type":"object","required":["labels","annotations","startsAt","endsAt","generatorURL"],"title":"AlertmanagerAlert"},"AlertmanagerWebhook":{"properties":{"version":{"type":"string","title":"Version"},"groupKey":{"type":"string","title":"Groupkey"},"truncatedAlerts":{"type":"integer","title":"Truncatedalerts"},"status":{"type":"string","title":"Status"},"receiver":{"type":"string","title":"Receiver"},"groupLabels":{"additionalProperties":{"type":"string"},"type":"object","title":"Grouplabels"},"commonLabels":{"additionalProperties":{"type":"string"},"type":"object","title":"Commonlabels"},"commonAnnotations":{"additionalProperties":{"type":"string"},"type":"object","title":"Commonannotations"},"externalURL":{"type":"string","title":"Externalurl"},"alerts":{"items":{"$ref":"#/components/schemas/AlertmanagerAlert"},"type":"array","title":"Alerts"}},"type":"object","required":["version","groupKey","truncatedAlerts","status","receiver","groupLabels","commonLabels","commonAnnotations","externalURL","alerts"],"title":"AlertmanagerWebhook"},"HTTPMethod":{"type":"string","enum":["CONNECT","DELETE","GET","HEAD","OPTIONS","PATCH","POST","PUT","TRACE"],"title":"HTTPMethod","description":"HTTP methods and descriptions\n\nMethods from the following RFCs are all observed:\n\n    * RFC 7231: Hypertext Transfer Protocol (HTTP/1.1), obsoletes 2616\n    * RFC 5789: PATCH Method for HTTP"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ICOSAppDeployedRequest":{"properties":{"app_descriptor":{"$ref":"#/components/schemas/ICOSAppDescriptor"},"app_instance":{"type":"string","title":"App Instance"},"common_action":{"$ref":"#/components/schemas/PolicyActionICOSService"},"service":{"type":"string","title":"Service"}},"type":"object","required":["app_descriptor","app_instance","common_action","service"],"title":"ICOSAppDeployedRequest"},"ICOSAppDescriptor":{"properties":{"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"},"components":{"items":{"$ref":"#/components/schemas/ICOSAppDescriptorComponent"},"type":"array","title":"Components","default":[]},"policies":{"items":{"$ref":"#/components/schemas/ICOSAppDescriptorPolicy"},"type":"array","title":"Policies","default":[]}},"type":"object","required":["name","description"],"title":"ICOSAppDescriptor"},"ICOSAppDescriptorComponent":{"properties":{"name":{"type":"string","title":"Name"},"type":{"type":"string","title":"Type"},"policies":{"anyOf":[{"items":{"$ref":"#/components/schemas/ICOSAppDescriptorPolicy"},"type":"array"},{"type":"null"}],"title":"Policies","default":[]}},"type":"object","required":["name","type"],"title":"ICOSAppDescriptorComponent"},"ICOSAppDescriptorPolicy":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"component":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Component"},"fromTemplate":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Fromtemplate"},"spec":{"anyOf":[{"$ref":"#/components/schemas/PolicySpecTemplate"},{"$ref":"#/components/schemas/PolicySpecTelemetry"},{"$ref":"#/components/schemas/PolicySpecConstraints"},{"type":"null"}],"title":"Spec"},"remediation":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Remediation"},"variables":{"type":"object","title":"Variables","default":{}},"properties":{"anyOf":[{"$ref":"#/components/schemas/PolicyProperties"},{"type":"null"}],"default":{}}},"type":"object","title":"ICOSAppDescriptorPolicy"},"PolicyActionICOSService":{"properties":{"type":{"const":"icos-service","title":"Type","default":"icos-service"},"uri":{"type":"string","title":"Uri"},"http_method":{"$ref":"#/components/schemas/HTTPMethod"},"extra_params":{"additionalProperties":{"type":"string"},"type":"object","title":"Extra Params","default":{}},"include_access_token":{"type":"boolean","title":"Include Access Token","default":false}},"type":"object","required":["uri","http_method"],"title":"PolicyActionICOSService"},"PolicyActionWebhook":{"properties":{"type":{"const":"webhook","title":"Type","default":"webhook"},"url":{"type":"string","title":"Url"},"httpMethod":{"$ref":"#/components/schemas/HTTPMethod"},"extraParams":{"additionalProperties":{"type":"string"},"type":"object","title":"Extraparams","default":{}},"includeAccessToken":{"type":"boolean","title":"Includeaccesstoken","default":false}},"type":"object","required":["url","httpMethod"],"title":"PolicyActionWebhook"},"PolicyCreate":{"properties":{"name":{"type":"string","title":"Name"},"subject":{"anyOf":[{"$ref":"#/components/schemas/PolicySubjectApplication"},{"$ref":"#/components/schemas/PolicySubjectHost"},{"$ref":"#/components/schemas/PolicySubjectCustom"}],"title":"Subject"},"spec":{"anyOf":[{"$ref":"#/components/schemas/PolicySpecTemplate"},{"$ref":"#/components/schemas/PolicySpecTelemetry"},{"$ref":"#/components/schemas/PolicySpecConstraints"}],"title":"Spec"},"action":{"$ref":"#/components/schemas/PolicyActionWebhook"},"variables":{"additionalProperties":{"anyOf":[{"type":"string"},{"type":"integer"},{"type":"number"}]},"type":"object","title":"Variables","default":{}},"properties":{"allOf":[{"$ref":"#/components/schemas/PolicyProperties"}],"default":{}}},"type":"object","required":["name","subject","spec","action"],"title":"PolicyCreate"},"PolicyEvent":{"properties":{"type":{"$ref":"#/components/schemas/PolicyEventType"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"details":{"additionalProperties":{"anyOf":[{"type":"string"},{"type":"integer"},{"type":"number"},{"additionalProperties":{"type":"string"},"type":"object"},{"type":"null"}]},"type":"object","title":"Details","default":{}}},"type":"object","required":["type","timestamp"],"title":"PolicyEvent"},"PolicyEventType":{"type":"string","enum":["activated","deactivated","violated","resolved","created","deleted"],"title":"PolicyEventType"},"PolicyPhase":{"type":"string","enum":["enforced","violated","inactive","unknown"],"title":"PolicyPhase"},"PolicyProperties":{"properties":{"oneoff":{"type":"boolean","title":"Oneoff"},"interval":{"type":"string","title":"Interval"},"pendingInterval":{"type":"string","title":"Pendinginterval"}},"type":"object","title":"PolicyProperties"},"PolicyRead":{"properties":{"name":{"type":"string","title":"Name"},"subject":{"anyOf":[{"$ref":"#/components/schemas/PolicySubjectApplication"},{"$ref":"#/components/schemas/PolicySubjectHost"},{"$ref":"#/components/schemas/PolicySubjectCustom"}],"title":"Subject"},"spec":{"anyOf":[{"$ref":"#/components/schemas/PolicySpecTemplate"},{"$ref":"#/components/schemas/PolicySpecTelemetry"},{"$ref":"#/components/schemas/PolicySpecConstraints"}],"title":"Spec"},"action":{"$ref":"#/components/schemas/PolicyActionWebhook"},"variables":{"additionalProperties":{"anyOf":[{"type":"string"},{"type":"integer"},{"type":"number"}]},"type":"object","title":"Variables","default":{}},"properties":{"allOf":[{"$ref":"#/components/schemas/PolicyProperties"}],"default":{}},"id":{"type":"string","title":"Id"},"status":{"allOf":[{"$ref":"#/components/schemas/PolicyStatus"}],"default":{"measurementBackends":{},"events":[],"phase":"unknown"}}},"type":"object","required":["name","subject","spec","action","id"],"title":"PolicyRead"},"PolicySpecConstraints":{"properties":{"description":{"type":"string","title":"Description","default":""},"type":{"const":"constraints","title":"Type"},"constraints":{"additionalProperties":{"type":"string"},"type":"object","title":"Constraints"}},"type":"object","required":["type","constraints"],"title":"PolicySpecConstraints"},"PolicySpecTelemetry":{"properties":{"description":{"type":"string","title":"Description","default":""},"type":{"const":"telemetryQuery","title":"Type","default":"telemetryQuery"},"expr":{"type":"string","title":"Expr"},"violatedIf":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Violatedif"},"thresholds":{"anyOf":[{"additionalProperties":{"type":"number"},"type":"object"},{"type":"null"}],"title":"Thresholds"}},"type":"object","required":["expr"],"title":"PolicySpecTelemetry"},"PolicySpecTemplate":{"properties":{"description":{"type":"string","title":"Description","default":""},"type":{"const":"template","title":"Type","default":"template"},"templateName":{"type":"string","title":"Templatename"}},"type":"object","required":["templateName"],"title":"PolicySpecTemplate"},"PolicyStatus":{"properties":{"measurementBackends":{"additionalProperties":{"additionalProperties":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"object"},"type":"object","title":"Measurementbackends","default":{}},"events":{"items":{"$ref":"#/components/schemas/PolicyEvent"},"type":"array","title":"Events","default":[]},"phase":{"allOf":[{"$ref":"#/components/schemas/PolicyPhase"}],"default":"unknown"}},"type":"object","title":"PolicyStatus"},"PolicySubjectApplication":{"properties":{"type":{"const":"app","title":"Type","default":"app"},"appName":{"type":"string","title":"Appname"},"appComponent":{"type":"string","title":"Appcomponent"},"appInstance":{"type":"string","title":"Appinstance"}},"type":"object","required":["appName","appComponent","appInstance"],"title":"PolicySubjectApplication"},"PolicySubjectCustom":{"properties":{"type":{"const":"custom","title":"Type","default":"custom"}},"additionalProperties":true,"type":"object","title":"PolicySubjectCustom"},"PolicySubjectHost":{"properties":{"type":{"const":"host","title":"Type","default":"host"},"hostId":{"type":"string","title":"Hostid"},"agentId":{"type":"string","title":"Agentid"}},"type":"object","required":["hostId","agentId"],"title":"PolicySubjectHost"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}